Refactor: move eirini to top level
